vm-allocator: Add SystemAllocator trait
The trait works for each vmm that has different resources and callback
and different allocation principle.

But from the view of crate vm-device, it doesn't know which kind of vmm and
resources are used, so try to allocate all resources supposed to have.
So SystemAllocator trait needs tell all kinds of allocator callback.

Signed-off-by: Jing Liu <jing2.liu@linux.intel.com>